Added the WizardDialog class.  The implementation of this class is not yet complete.
git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2200 897c5806-f142-0410-a293-e1bf5698dec8

+++ org.gamegineer.common.ui/src/org/gamegineer/common/ui/dialog/AbstractDialog.java
@@ -291,7 +291,7 @@ public abstract class AbstractDialog
         createButtonsForButtonBar( composite );
 
         final GroupLayout.Group hGroup = layout.createSequentialGroup();
-        hGroup.addGap( 0, 0, Short.MAX_VALUE );
+        hGroup.addGap( 0, 0, Integer.MAX_VALUE );
         final GroupLayout.Group vGroup = layout.createParallelGroup();
         final int hGap = convertWidthInDlusToPixels( DialogConstants.HORIZONTAL_SPACING );
         int index = 0;
